+++ MobileApp/app/src/main/java/pt/ipleiria/estg/dei/waypinpoint/utils/Utilities.java
@@ -41,7 +41,7 @@ public class Utilities {
     public static final String TOKEN = "TOKEN";
     public static final String SNACKBAR_MESSAGE = "SNACKBAR_MESSAGE";
     public static final String ID_CART = "ID_CART";
-    public static final int DB_VERSION = 3;
+    public static final int DB_VERSION = 4;
